+++ src/main/kotlin/shop/itbug/fluttercheckversionx/actions/riverpod/RiverpodAutoImportTextConfig.kt
@@ -17,7 +17,7 @@ class RiverpodAutoImportTextConfig : AnAction() {
             val state = PluginConfig.getState(project)
             WidgetUtil.getTextEditorPopup(
                 title,
-                state.autoImportRiverpodText,
+                state.autoImportRiverpodText ?: "",
                 state.autoImportRiverpodText,
                 null, {
                     it.showCenteredInCurrentWindow(project)