+++ src/main/kotlin/shop/itbug/fluttercheckversionx/window/preview/ImagesPreviewWindow.kt
@@ -1,26 +1,31 @@
 package shop.itbug.fluttercheckversionx.window.preview
 
 import com.intellij.ide.actions.CopyReferencePopup
-import com.intellij.openapi.actionSystem.ActionManager
-import com.intellij.openapi.actionSystem.CommonDataKeys
-import com.intellij.openapi.actionSystem.DataSink
-import com.intellij.openapi.actionSystem.UiDataProvider
+import com.intellij.openapi.Disposable
+import com.intellij.openapi.actionSystem.*
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.project.guessProjectDir
+import com.intellij.openapi.util.text.StringUtil
+import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.VirtualFileVisitor
+import com.intellij.openapi.wm.ToolWindow
 import com.intellij.ui.JBColor
 import com.intellij.ui.PopupHandler
 import com.intellij.ui.components.JBLabel
 import com.intellij.ui.components.JBPanel
 import com.intellij.ui.scale.JBUIScale
+import com.intellij.util.Alarm
 import com.intellij.util.ui.JBUI
 import com.intellij.util.ui.UIUtil
 import com.intellij.util.ui.WrapLayout
 import com.intellij.util.ui.components.BorderLayoutPanel
+import shop.itbug.fluttercheckversionx.actions.imagePreview.toolbar.FlutterAssetsPreviewPanelToolbarActionGroup
+import shop.itbug.fluttercheckversionx.actions.imagePreview.toolbar.PreviewSearchTextField
 import shop.itbug.fluttercheckversionx.common.scroll
+import shop.itbug.fluttercheckversionx.config.PluginConfig
+import shop.itbug.fluttercheckversionx.i18n.PluginBundle
 import shop.itbug.fluttercheckversionx.util.ImageFileUtil
 import shop.itbug.fluttercheckversionx.util.calculateAspectRatio
 import java.awt.Container
@@ -28,33 +33,94 @@ import java.awt.Cursor
 import java.awt.Dimension
 import java.awt.event.MouseAdapter
 import java.awt.event.MouseEvent
-import javax.swing.BorderFactory
-import javax.swing.JComponent
-import javax.swing.SwingConstants
-import javax.swing.SwingUtilities
+import javax.swing.*
 import javax.swing.event.PopupMenuEvent
 import javax.swing.event.PopupMenuListener
 
 /**
  * 资产图片预览窗口
  */
-class ImagesPreviewWindow(val project: Project) : BorderLayoutPanel() {
+class ImagesPreviewWindow(val project: Project, val toolWindow: ToolWindow) : BorderLayoutPanel(), UiDataProvider,
+    Disposable {
 
-
+    val pluginConfig = PluginConfig.getState(project)
     private val assetsDirector: VirtualFile?
         get() = ApplicationManager.getApplication().executeOnPooledThread<VirtualFile?> {
-            return@executeOnPooledThread project.guessProjectDir()?.findChild("assets")
+            if (pluginConfig.assetDirectory == null) {
+                return@executeOnPooledThread null
+            }
+            return@executeOnPooledThread LocalFileSystem.getInstance().findFileByPath(pluginConfig.assetDirectory!!)
         }.get()
     private val visitor = FileVisitor()
     val panel = ImagePanel()
 
+    val searchField = PreviewSearchTextField(onSearchChanged)
+    val fileComps = mutableSetOf<AssetFileLayout>()
+
+
+    val toolbar = ActionManager.getInstance().createActionToolbar(
+        "Assets Preview Toolbar",
+        FlutterAssetsPreviewPanelToolbarActionGroup.getActionGroup(), true
+    ).apply {
+        targetComponent = toolWindow.component
+    }
+
+    val scp = panel.scroll()
+    val alarm = Alarm(this)
+
+
+    val onSearchChanged: (String) -> Unit
+        get() = { str ->
+            alarm.cancelAllRequests()
+            alarm.addRequest({
+                panel.removeAll()
+                if (str.isNotEmpty()) {
+                    fileComps.forEach {
+                        val fileName = it.file.nameWithoutExtension
+                        if (StringUtil.toLowerCase(fileName).contains(StringUtil.toLowerCase(str))) {
+                            panel.add(it)
+                        }
+                    }
+                } else {
+                    fileComps.map(panel::add)
+                }
+                ApplicationManager.getApplication().invokeLater {
+                    alarm.addRequest({
+                        panel.repaint()
+                        panel.updateUI()
+                    }, 334)
+                }
+            }, 330)
+
+        }
+
     init {
 
-        addToCenter(panel.scroll().apply {
-            border = JBUI.Borders.empty(12)
+        addToTop(BorderLayoutPanel().apply {
+            addToCenter(searchField)
+            addToRight(toolbar.component)
         })
+        addToCenter(scp)
+
         SwingUtilities.invokeLater {
-            startLoadAssets()
+            if (pluginConfig.enableAssetsPreviewAction) {
+                startLoadAssets()
+            } else {
+                addToCenter(JBLabel(PluginBundle.get("action_not_open_and_enable")).apply {
+                    horizontalAlignment = SwingConstants.CENTER
+                })
+            }
+        }
+
+        addNullDirectoryTips()
+    }
+
+    private fun addNullDirectoryTips() {
+        val dir = pluginConfig.assetDirectory
+        if (dir.isNullOrBlank()) {
+            addToCenter(JBLabel("Assets Preview Directory Not Found\nPlease set in the settings").apply {
+                horizontalAlignment = SwingConstants.CENTER
+            })
         }
     }
 
@@ -72,6 +138,24 @@ class ImagesPreviewWindow(val project: Project) : BorderLayoutPanel() {
 
     }
 
+    override fun uiDataSnapshot(sink: DataSink) {
+        sink[KEY] = this
+    }
+
+    companion object {
+        val KEY = DataKey.create<ImagesPreviewWindow>("FlutterXAssetsPreviewWindow")
+    }
+
+
+    fun refreshItems() {
+        panel.removeAll()
+        startLoadAssets()
+        println("刷新")
+    }
+
+    override fun dispose() {
+
+    }
 
     // 访问图片
     inner class FileVisitor() : VirtualFileVisitor<VirtualFile>() {
@@ -81,6 +165,7 @@ class ImagesPreviewWindow(val project: Project) : BorderLayoutPanel() {
             }
             return CONTINUE
         }
+
     }
 
     // 面板 ui
@@ -88,28 +173,39 @@ class ImagesPreviewWindow(val project: Project) : BorderLayoutPanel() {
 
         //添加图片资产
         fun addImageAsset(asset: VirtualFile) {
-            add(AssetFileLayout(project, asset))
+            val item = AssetFileLayout(project, asset)
+            fileComps.add(item)
+            add(item)
         }
     }
 }
 
-private class AssetFileLayout(project: Project, val file: VirtualFile) : BorderLayoutPanel(), UiDataProvider {
-    val image = ImageFileUtil.getIcon(file)
+class AssetFileLayout(project: Project, val file: VirtualFile) : BorderLayoutPanel(), UiDataProvider {
+    val pluginConfig = PluginConfig.getState(project)
+    val image: ImageIcon
+        get() = ApplicationManager.getApplication().executeOnPooledThread<ImageIcon> { ImageFileUtil.getIcon(file) }
+            .get()
+    val imageItemSize = pluginConfig.assetsPreviewImageSize
+    val resizeIcon = ImageFileUtil.resizeImageIconCover(image, imageItemSize, imageItemSize)
     val imageSize = ImageFileUtil.getSize(file, project)
     val nameLabel = JBLabel(
         "${file.name} (${imageSize.width}x${imageSize.height},${
             calculateAspectRatio(
                 imageSize.width, imageSize.height
             )
-        })", UIUtil.ComponentStyle.MINI
+        })", UIUtil.ComponentStyle.SMALL
     ).apply {
         horizontalAlignment = SwingConstants.CENTER
         border = JBUI.Borders.empty(5, 0)
     }
     private var highlightEnabled = false
 
+    private val imageLabel = JBLabel(resizeIcon).apply {
+
+    }
+
     init {
-        addToCenter(JBLabel(image))
+        addToCenter(imageLabel)
         addToBottom(nameLabel)
         installPopup(this)
         addMouseListener(object : MouseAdapter() {